Autodelete, more languages
Java and C++ are working.
If successive runs are made, the previous run's temp file(s) are deleted before the new extension is detected. The last run of a session (before a reload of Obsidian) won't be deleted automatically, though, so the Clear command should still be used.
A test is run to verify the selection's validity before running the rest of the code, preventing Obsidian from freezing in an infinite loop searching fro a char that doesn't exist.
Comments have been added for code legibility.